+++ collect-webapp/frontend/src/datamanagement/components/recordeditor/fields/CompositeAttributeFormItem.js
@@ -1,13 +1,16 @@
 import React from 'react'
 import { Col, Label, Row } from 'reactstrap'
 
-const CompositeAttributeFormItem = ({ field, label, inputField, labelWidth = 50 }) => (
-  <Row key={field}>
-    <Col style={{ maxWidth: `${labelWidth}px` }}>
-      <Label>{label}</Label>
-    </Col>
-    <Col>{inputField}</Col>
-  </Row>
-)
+const CompositeAttributeFormItem = ({ field, label, inputField, labelWidth = 50 }) => {
+  const widthPx = `${labelWidth}px`
+  return (
+    <Row key={field}>
+      <Col style={{ width: widthPx, maxWidth: widthPx }}>
+        <Label>{label}</Label>
+      </Col>
+      <Col>{inputField}</Col>
+    </Row>
+  )
+}
 
 export default CompositeAttributeFormItem